Assignment 3 complete (except for portions of extra credit #2.

+++ Assignment_3_Set_iOS9/Set/SetCard.m
@@ -57,7 +57,7 @@
 
 - (NSString *)description
 {
-  return [NSString stringWithFormat:@"%lu-%@-%@-%@", self.number, self.shade, [self.shape description], [self.color description]];
+  return [NSString stringWithFormat:@"%lu-%@-%@-%@", (unsigned long)self.number, self.shade, [self.shape description], [self.color description]];
 }
 
 #pragma mark - Class Methods
